+++ src/Faker.Tests/InternetTests.cs
@@ -10,6 +10,20 @@ namespace RimuTec.Faker.Tests {
    [TestFixture]
    public class InternetTests {
       [Test]
+      public void DomainSuffix_HappyDays() {
+         // arrange
+
+         // act
+         var suffix = Internet.DomainSuffix();
+
+         // assert
+         Assert.IsFalse(string.IsNullOrWhiteSpace(suffix));
+         Assert.IsFalse(suffix.Contains("#"));
+         Assert.IsFalse(suffix.Contains("?"));
+         Assert.AreEqual(0, Regex.Matches(suffix, @"[A-Z]").Count);
+      }
+
+      [Test]
       public void DomainWord_HappyDays() {
          // arrange