+++ releasenotes.md
@@ -2,6 +2,7 @@
 
 ## Version 0.17.0
 ### New Features
+- Internet.DomainSuffix()
 - Internet.DomainWord()
 
 ## Version 0.16.0